Fix issue with requests missing PI information.
The original resource actions staging query pulled in the resource actions
which had a PI associated with them. Turns out that there were two different
requests that had no PIO assigned a the time of request (one case the PI was
added a day later and the other case 2 weeks later). This changes the query to
a dependent subquery that selects the PI that was on the project closest in
time to the request.